yatube/posts/tests/test_models.py
from django.test import TestCase

from ..models import User, Group, Post, Comment, Follow


class PostModelTest(TestCase):
    @classmethod
    def setUpClass(cls):
        super().setUpClass()
        cls.user = User.objects.create_user(username='auth')
        cls.user_2 = User.objects.create_user(username='auth_2')
        cls.group = Group.objects.create(
            title='Тестовая группа',
            slug='Тестовый слаг',
            description='Тестовое описание',
        )
        cls.post = Post.objects.create(
            author=cls.user,
            group=cls.group,
            text='Тестовый пост',
        )
        cls.comment = Comment.objects.create(
            post=cls.post,
            author=cls.user,
            text='Тестовый комментарий',
        )
        cls.subscription = Follow.objects.create(
            user=cls.user_2,
            author=cls.user
        )

    def test_models_have_correct_object_names(self):
        """Проверяем, что у моделей корректно работает __str__."""
        self.assertEqual(self.post.text[:15], str(self.post))
        self.assertEqual(self.comment.text[:15], str(self.comment))
        self.assertEqual(self.post.group.title, str(self.post.group))
        self.assertEqual(
            f'{self.subscription.user.username}-'
            f'{self.subscription.author.username}',
            str(self.subscription))
